Servicio para agregar filtros a partir de campos
